Get rid of StringDiagram
