+++ src/diagram/chord_diagram.rs
@@ -1,6 +1,5 @@
 use crate::chord::FretID;
 use crate::diagram::FretPattern;
-use crate::diagram::StringDiagram;
 use crate::diagram::CHART_WIDTH;
 use crate::note::Note;
 use crate::STRING_COUNT;
@@ -41,6 +40,44 @@ impl ChordDiagram {
     }
 }
 
+/// Format a line that represents a ukulele string in a chord diagram.
+fn format_line(
+    root: Note,
+    base_fret: FretID,
+    fret: FretID,
+    note: Note,
+    root_width: usize,
+) -> String {
+    let root_str = format!("{:width$}", root.to_string(), width = root_width);
+
+    // Show a symbol for the nut if the chord is played on the lower
+    // end of the fretboard. Indicate ongoing strings otherwise.
+    let nut = match base_fret {
+        1 => "||",
+        _ => "-|",
+    };
+
+    // Mark open strings with a special symbol.
+    let sym = match fret {
+        0 => "o",
+        _ => " ",
+    };
+
+    // Create a line representing the string with the fret to be pressed.
+    let mut string = "".to_owned();
+
+    for i in base_fret..base_fret + CHART_WIDTH {
+        let c = match fret {
+            fret if fret == i => "o",
+            _ => "-",
+        };
+
+        string.push_str(&format!("-{}-|", c));
+    }
+
+    format!("{} {}{}{}- {}", root_str, sym, nut, string, note)
+}
+
 impl fmt::Display for ChordDiagram {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         let mut s = "".to_string();
@@ -59,8 +96,9 @@ impl fmt::Display for ChordDiagram {
 
         // Create a diagram for each ukulele string.
         for (root, fret, note) in izip!(&self.roots, self.frets.iter(), &self.notes).rev() {
-            let sd = StringDiagram::new(*root, base_fret, *fret, *note, root_width);
-            s.push_str(&format!("{}\n", sd.to_string()));
+            let sd = format_line(*root, base_fret, *fret, *note, root_width);
+            s.push_str(&sd);
+            s.push('\n');
         }
 
         // If the fretboard section shown does not include the nut,
@@ -79,6 +117,42 @@ mod tests {
     use indoc::indoc;
     use rstest::rstest;
     use std::str::FromStr;
+
+    #[rstest(
+        root_name,
+        base_fret,
+        fret,
+        note_name,
+        diagram,
+        case("C", 1, 0, "C", "C o||---|---|---|---|- C"),
+        case("C", 1, 4, "E", "C  ||---|---|---|-o-|- E"),
+        case("C", 1, 2, "D", "C  ||---|-o-|---|---|- D"),
+        case("G", 1, 4, "B", "G  ||---|---|---|-o-|- B"),
+        case("C", 5, 7, "G", "C  -|---|---|-o-|---|- G"),
+        case("C", 1, 1, "C#", "C  ||-o-|---|---|---|- C#"),
+        case("C", 1, 1, "Db", "C  ||-o-|---|---|---|- Db"),
+        case("C", 5, 6, "F#", "C  -|---|-o-|---|---|- F#"),
+        case("C", 5, 6, "Gb", "C  -|---|-o-|---|---|- Gb"),
+        case("F#", 1, 0, "F#", "F# o||---|---|---|---|- F#"),
+        case("F#", 1, 4, "A#", "F#  ||---|---|---|-o-|- A#"),
+        case("F#", 5, 7, "D", "F#  -|---|---|-o-|---|- D")
+    )]
+    fn test_format_line(
+        root_name: &str,
+        base_fret: FretID,
+        fret: FretID,
+        note_name: &str,
+        diagram: &str,
+    ) {
+        let root = Note::from_str(root_name).unwrap();
+        let note = Note::from_str(note_name).unwrap();
+
+        let root_width = root_name.len();
+
+        let sd = format_line(root, base_fret, fret, note, root_width);
+
+        assert_eq!(sd, diagram);
+    }
 
     #[rstest(chord_name, min_fret, tuning, diagram,
         case(
